Don't mark all lanes as designated when there are blank lane specifiers in the tag.
