+++ src/tools/extract.cpp
@@ -1,6 +1,6 @@
 #include "extractor/extractor.hpp"
 #include "extractor/extractor_config.hpp"
-#include "extractor/scripting_environment_lua.hpp"
+#include "extractor/scripting_environment_sol2.hpp"
 #include "util/log.hpp"
 #include "util/version.hpp"
 
@@ -153,7 +153,7 @@ int main(int argc, char *argv[]) try
     }
 
     // setup scripting environment
-    extractor::LuaScriptingEnvironment scripting_environment(
+    extractor::Sol2ScriptingEnvironment scripting_environment(
         extractor_config.profile_path.string().c_str());
     auto exitcode = extractor::Extractor(extractor_config).run(scripting_environment);
